[Patchset 9] Remove warning suppressions and fix many casting bugs
This change removes most warning suppressions, in particular
from production code. In the past we inherit a bunch of -Wno-xxx
suppressions required to build gtest and libprotobuf headers.
Doing so, however, caused the suppressions to propagate back to
the translation units that were including any protobuf header
or any auto-generated .pb.h stub.
This change moves the gtest and probobuf header to be a
system include (-isystem vs -I). Doing so implicitly blacklists
any compiler warning on the headers.
This CL then re-enables warnings and deals with the fall out of
fixes that came out of this.

Bug: 77316877
Test: pefetto_unittests / perfetto_integrationtests

+++ src/ftrace_reader/ftrace_controller_unittest.cc
@@ -79,7 +79,7 @@ class MockTaskRunner : public base::TaskRunner {
   }
 
   MOCK_METHOD1(PostTask, void(std::function<void()>));
-  MOCK_METHOD2(PostDelayedTask, void(std::function<void()>, int delay_ms));
+  MOCK_METHOD2(PostDelayedTask, void(std::function<void()>, uint32_t delay_ms));
   MOCK_METHOD2(AddFileDescriptorWatch, void(int fd, std::function<void()>));
   MOCK_METHOD1(RemoveFileDescriptorWatch, void(int fd));
 
@@ -430,9 +430,8 @@ TEST(FtraceControllerTest, DISABLED_DrainPeriodRespected) {
   std::unique_ptr<FtraceSink> sink = controller->CreateSink(config, &delegate);
 
   const int kCycles = 50;
-  EXPECT_CALL(
-      *controller->runner(),
-      PostDelayedTask(_, static_cast<int>(controller->drain_period_ms())))
+  EXPECT_CALL(*controller->runner(),
+              PostDelayedTask(_, controller->drain_period_ms()))
       .Times(kCycles);
   EXPECT_CALL(*controller, OnRawFtraceDataAvailable(_)).Times(kCycles);
   EXPECT_CALL(*controller->runner(), PostTask(_)).Times(kCycles);
