+++ src/perfetto_cmd/perfetto_cmd.cc
@@ -259,9 +259,8 @@ void PerfettoCmd::OnConnect() {
 
   // Failsafe mechanism to avoid waiting indefinitely if the service hangs.
   if (trace_config_->duration_ms()) {
-    task_runner_.PostDelayedTask(
-        std::bind(&PerfettoCmd::OnTimeout, this),
-        static_cast<int>(trace_config_->duration_ms() + 10000));
+    task_runner_.PostDelayedTask(std::bind(&PerfettoCmd::OnTimeout, this),
+                                 trace_config_->duration_ms() + 10000);
   }
 }
